добавил TestBase класс для оптимизации проекта теперь все общие функции находятся именно в этом классе
