+++ addressbook-web-tests/src/test/java/ru/stqa/qajava/addressbook/tests/GroupDeletionTests.java
@@ -1,0 +1,16 @@
+package ru.stqa.qajava.addressbook.tests;
+
+import org.testng.annotations.Test;
+
+public class GroupDeletionTests extends TestBase {
+
+
+  @Test
+  public void testGroupDeletion() throws Exception {
+    gotoGroupPage();
+    selectGroup();
+    deleteSelectedGroups();
+    reternToGroupPage();
+  }
+
+}
